add unitplaceholders, unit attacks, unit placeable
added unit class with properties and is able to attack mobs.
added so mobs can die when attacked and the wave starts correctly when all
mobs is dead. Also made so units can be placed on nodes.
Units also updates the target and the attackprojectile seeks the target
until hit.